requirements.txt
coinbase-advanced-py
ccxt
xgboost>=3
torch
optuna
ta
pandas
numpy
scikit-learn
sqlalchemy
fastapi
uvicorn[standard]
websockets
click
rich
pyyaml
python-dotenv
pytest
pytest-asyncio
aiohttp
httpx
